+++ src/main/java/com/levelUpToast/levelUpToast/rest/mail/MailController.java
@@ -3,7 +3,7 @@ package com.levelUpToast.levelUpToast.rest.mail;
 import com.levelUpToast.levelUpToast.config.exception.LevelUpToastEx;
 import com.levelUpToast.levelUpToast.domain.dataForm.mail.MailResponseForm;
 import com.levelUpToast.levelUpToast.domain.mail.Mail;
-import com.levelUpToast.levelUpToast.service.e_mail.mailServiceInf.MailService;
+import com.levelUpToast.levelUpToast.service.mail.mailServiceInf.MailService;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
